Ticket #860rzm1k4 ListadoAlojamiento: cuando se filtra que no muestre todas las habitaciones del hotel, solo de la fecha seleccionada

+++ TPCAI/Entidades/SubClasses/HabitacionesHotelSubClass.cs
@@ -8,7 +8,7 @@ namespace TPCAI
 {
     public class HabitacionesHotelSubClass
     {
-        public long IDHabitacion { get; set; }
+        public int IDHabitacion { get; set; }
         public DateTime FechaHabitacionHotel { get; set; }
         public int Cantidad { get; set; }
